file organization, range_value working (default value LONG), added stop method, documentation changes
