+++ robotic_skin/sensor/vl53l1x.py
@@ -2,18 +2,16 @@
 This is a python class for VL53L1X Proximity Sensor
 Datasheet Link: https://www.st.com/resource/en/datasheet/vl53l1x.pdf
 This library heavily uses VL53L1X pip application: pip install VL53L1X
-TODO: Find why the application is running SLOW
 """
 
 import VL53L1X
 from robotic_skin.sensor import Sensor
 
-
 class VL53L1X_ProximitySensor(Sensor):
     """
     Code for VL53L1X distance sensor class.
     """
-    def __init__(self, i2c_bus=1, i2c_address=0x29, range_value=3):
+    def __init__(self, i2c_bus=1, i2c_address=0x29, range_value=3, timing_budget=100000, inter_measurement_period = 100):
         """
         Initialize the VL53L1X sensor, test if the python code can reach it or not, if not throw an exception
         Parameters
@@ -27,27 +25,28 @@ class VL53L1X_ProximitySensor(Sensor):
             note this can be changed to anything if you wish so. This library doesn't handle changing I2C address as of
             now
         range_value : int
-            The proximity sensor has 3 ranges, according to the Python Library. This is exactly that int.
+            The proximity sensor has 3 ranges, according to the Python Library:
+                NONE = 0
+                SHORT = 1
+                MEDIUM = 2
+                LONG = 3
             Link: https://pypi.org/project/VL53L1X/
             By default it's kept to long range
+        timing_budget : int
+            Timing budget in microseconds. 
+            A higher timing budget results in greater measurement accuracy, but also a higher power consumption.
+        inter_measurement_period : int
+            Inter measurement period in miliseconds. 
+            The inter measurement period must be >= the timing budget, otherwise it will be double the expected value.
         """
         super().__init__()
         self.tof = VL53L1X.VL53L1X(i2c_bus, i2c_address)
         self.tof.open()
-        if range_value in (1, 2, 3):
+        if range_value in (0, 1, 2, 3):
             self.tof.start_ranging(range_value)
-            self.tof.set_timing(22000, 100)
+            self.tof.set_timing(timing_budget, inter_measurement_period)
         else:
-            raise Exception("The range value passed is not 1 or 2 or 3")
-
-    def calibrate(self):
-        """
-        This is the calibration function.
-        # TODO: Decide whether you have to implement it or not
-        Returns
-        -------
-        None
-        """
+            raise "The range value passed is not 0 or 1 or 2 or 3"
 
     def _read_raw(self):
         """
@@ -72,6 +71,7 @@ class VL53L1X_ProximitySensor(Sensor):
         float
             Corrected value from raw value
         """
+        #TODO 
         return input_value
 
     def read(self):
@@ -86,11 +86,8 @@ class VL53L1X_ProximitySensor(Sensor):
         """
         return self._calibrate_values(self._read_raw())
 
-
-if __name__ == "__main__":
-    # Test case
-    from time import sleep
-    ps = VL53L1X_ProximitySensor()
-    while True:
-        print("Proximity Sensor Reading: ", ps.read())
-        sleep(0.02)
+    def stop(self):
+        """
+        Stop VL53L1X ToF Sensor Ranging
+        """
+        self.tof.stop_ranging()
